# Improve performance of PurgeCSS extractor result merging