--- a/packages/purgecss/package.json
+++ b/packages/purgecss/package.json
@@ -45,7 +45,8 @@
     "postcss-selector-parser": "^6.0.2"
   },
   "devDependencies": {
-    "@types/glob": "^7.1.1"
+    "@types/glob": "^7.1.1",
+    "tailwindcss": "^1.2.0"
   },
   "bugs": {
     "url": "https://github.com/FullHuman/purgecss/issues"