## Proposed changes

### Add a performance regression test to `packages/purgecss` (fa54f1194cd35efb60ec5b517c2715cfa5573838)

This test uses all of the `.js` files in `node_modules` to purge all of the `.css` files in `__tests__/test_examples`, including the newly added [`tailwind.css`](https://tailwindcss.com/), which is a whopping 908KB of CSS (before purging).

On my machine, this all takes about 50 seconds, though it may exceed the 60sec `jest.setTimeout` when running in a CI environment. I will push additional commits that improve performance significantly.

Although this test may seem extreme, the Tailwind docs [recommend](https://tailwindcss.com/docs/controlling-file-size/#setting-up-purgecss) a PurgeCSS setup that can easily resemble (or dwarf) this test, when used in larger applications with lots of content files and lots of CSS files.

### Avoid expensive copies of `ExtractorResultDetailed` objects (5c5bcdcd83e0e0d4aa29c2f5e63aa3647f9e376c)

Should fix #300.

First, I created an `ExtractorResultSets` class as the internal representation of `ExtractorResultDetailed` input objects, so that I could make its properties private and query them with meaningful methods. This allowed me to safely change the types used to represent the internal collections of strings from `string[]` to `Set<string>`, for faster merging and lookup.

Thanks to the new `ExtractorResultSets.merge` method, which modifies its receiver object (`this`) rather than returning a new `ExtractorResultSets` object, we no longer have to call `mergeExtractorSelectors` repeatedly to create unnecessary intermediate copies of `ExtractorResultDetailed` objects.

Avoiding these expensive copies can be a big win when the `ExtractorResultSets` objects contain many (tens or hundreds of thousands of) strings, saving those `Set<string>`s from getting repeatedly copied and then discarded as garbage memory. Instead, each new `ExtractorResultSets` object gets merged into a single `ExtractorResultSets` result object, leading to a lot less copying of strings and objects and `Set`s.

> Note: the `new Set([...oldSet, ...newValues])` pattern used by `purgecss@1.x` was wasteful in a similar way to `mergeExtractorSelectors`, because it required iterating over the entire `oldSet`, then concatenating its elements with the `newValues`, then creating a completely new `Set` out of that array. Writing code in this style tends to undo the performance benefits of using a `Set`.

## Types of changes

- [x] Performance regression tests
- [x] Bugfix (#300)

## Checklist

- [x] Lint and unit tests pass locally with my changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)