@benjamn 2020-04-05
Ok, if you drill into the test results, the time taken by the performance test dropped from 87476ms to 2029ms for Node.js 10 (-97.7%), and from 60777ms to 2387ms for Node.js 12 (-96.7%). Not bad for a Saturday afternoon of hacking around! :feelsgood:

After some additional profiling, I suspect the next area worth focusing on is `postcss-selector-parser`, which creates a relatively complicated `Parser` object for every CSS selector that gets processed. That's not something I have time to work on, but I believe @leeoniya was able to improve CSS parsing performance in [dropcss](https://github.com/leeoniya/dropcss): https://github.com/FullHuman/purgecss/issues/300#issuecomment-609071320