@leeoniya 2020-04-05
> @leeoniya was able to improve CSS parsing performance in dropcss: #300 (comment)

the major caveat(s) being that dropcss's parsers are intolerant of malformed markup and css, do not handle all escapes that can exist in valid code [1], cannot support sourcemaps, and do not provide helpful error messages.

even one of these drawbacks easily disqualifies them from many applications that must be more-or-less bomb-proof, and makes any perf comparison potentially useless, given the rest of the project's requirements.

[1] https://github.com/leeoniya/dropcss#special--escaped-sequences