@Ffloriel 2020-04-05
Thanks a lot for the explanations and for this amazing PR!

Not using custom parsers is on purpose. I was looking to limit the scope of what the project will have to maintain. With new CSS features, and pseudo-selectors coming along, `postcss` and `postcss-selector-parser` should follow quickly. It's obviously a tradeoff and if it could improve the performance a lot, replacing `postcss-selector-parser` might be worth it.