@benjamn 2020-04-05
@Ffloriel oh sorry I didn’t mean replacing `postcss-selector-parser`, just focusing some future performance attention on it, since it’s a hotspot for `purgecss`