@benjamn 2020-04-06
@Ffloriel I believe this is ready to merge, unless you notice anything else I should adjust. Please feel free to squash-and-merge rather than preserving the individual commits. Thanks for maintaining such an easy-to-work-with repository!